Need to define the PY_SSIZE_T_CLEAN macro in AEC.c for Python version 3.10. Changed from int to Py_ssize_t for some variables after the macro definition, or the cipher mode would always set to zero somehow.

+++ charm/core/crypto/cryptobase/block_template.h
@@ -46,7 +46,7 @@
 typedef struct
 {
 	PyObject_HEAD
-	int mode, count, segment_size, prf_mode;
+	Py_ssize_t mode, count, segment_size, prf_mode;
 	unsigned char IV[BLOCK_SIZE], oldCipher[BLOCK_SIZE];
 	PyObject *counter;
 	int counter_shortcut;
